+++ cardinal-components-base/src/main/java/dev/onyxstudios/cca/internal/base/asm/StaticComponentPluginBase.java
@@ -26,8 +26,10 @@ import dev.onyxstudios.cca.api.v3.component.Component;
 import dev.onyxstudios.cca.api.v3.component.ComponentContainer;
 import dev.onyxstudios.cca.api.v3.component.ComponentFactory;
 import dev.onyxstudios.cca.api.v3.component.ComponentKey;
+import dev.onyxstudios.cca.internal.base.ComponentRegistrationInitializer;
 import dev.onyxstudios.cca.internal.base.LazyDispatcher;
 import net.fabricmc.fabric.api.event.Event;
+import net.fabricmc.loader.api.FabricLoader;
 import net.fabricmc.loader.api.entrypoint.EntrypointContainer;
 import net.fabricmc.loader.api.metadata.ModMetadata;
 import org.objectweb.asm.MethodVisitor;
@@ -38,6 +40,7 @@ import org.objectweb.asm.tree.ClassNode;
 import java.io.IOException;
 import java.lang.reflect.Constructor;
 import java.lang.reflect.Method;
+import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collection;
 import java.util.Collections;
@@ -154,6 +157,18 @@ public abstract class StaticComponentPluginBase<T, I> extends LazyDispatcher {
                 throw new StaticComponentLoadingException(String.format("Exception while registering static component factories for %s (%s)", metadata.getName(), metadata.getId()), e);
             }
         }
+    }
+
+    public static <I extends ComponentRegistrationInitializer> Collection<EntrypointContainer<I>> getComponentEntrypoints(String key, Class<I> type) {
+        Collection<EntrypointContainer<ComponentRegistrationInitializer>> generic = FabricLoader.getInstance().getEntrypointContainers("cardinal-components", ComponentRegistrationInitializer.class);
+        Collection<EntrypointContainer<I>> specific = new ArrayList<>(FabricLoader.getInstance().getEntrypointContainers(key, type));
+        for (EntrypointContainer<ComponentRegistrationInitializer> container : generic) {
+            if (type.isInstance(container.getEntrypoint())) {
+                @SuppressWarnings("unchecked") EntrypointContainer<I> c = (EntrypointContainer<I>) container;
+                specific.add(c);
+            }
+        }
+        return specific;
     }
 
     protected abstract Collection<EntrypointContainer<I>> getEntrypoints();