+++ cardinal-components-scoreboard/src/main/java/dev/onyxstudios/cca/api/v3/scoreboard/ScoreboardComponentInitializer.java
@@ -22,18 +22,19 @@
  */
 package dev.onyxstudios.cca.api.v3.scoreboard;
 
+import dev.onyxstudios.cca.internal.base.ComponentRegistrationInitializer;
 import org.jetbrains.annotations.ApiStatus;
 
 /**
  * Entrypoint getting invoked to register <em>static</em> team component factories.
  *
- * <p>The entrypoint is exposed as {@code cardinal-components-scoreboard} in the mod json and runs for any environment.
+ * <p>The entrypoint is exposed as either {@code "cardinal-components"} or {@code "cardinal-components-scoreboard"} in the mod json and runs for any environment.
  * It usually executes right before the first {@link net.minecraft.scoreboard.Scoreboard} instance is created.
  *
  * @since 2.4.2
  */
 @ApiStatus.Experimental
-public interface ScoreboardComponentInitializer {
+public interface ScoreboardComponentInitializer extends ComponentRegistrationInitializer {
     /**
      * Called to register component factories for statically declared component types.
      *