+++ changelog.md
@@ -14,6 +14,13 @@ Changes
 - Replaced specific factory interfaces with a generic variant (eg. `ComponentFactory`)
 
 ------------------------------------------------------
+Version 2.7.5
+------------------------------------------------------
+Additions
+- Added the universal `cardinal-components` entrypoint key
+    - This new key can be used for any CCA registration initializer (replaces eg. `cardinal-components-entity`)
+
+------------------------------------------------------
 Version 2.7.4
 ------------------------------------------------------
 Additions